feat: Add Giza style
User is now able to manipulate the styles of the active view.
This is achieved by auto turning the sun and shadow intensity.
For these settings to apply, the User has to turn on the shadow
manually. The style is hardcoded in the Snap Command
but can be shifted to be an independent command.